moved logged_in to own folder, added to dos
